版本同步

+++ plugins/alipay/index.js
@@ -38,9 +38,25 @@ const createOrder = async (user, account, amount, orderType = 3) => {
   }
   const orderId = moment().format('YYYYMMDDHHmmss') + Math.random().toString().substr(2, 6);
   const time = 60;
+  const orderSetting = await knex('webguiSetting').select().where({
+    key: 'payment',
+  }).then(success => {
+    if(!success.length) {
+      return Promise.reject('settings not found');
+    }
+    success[0].value = JSON.parse(success[0].value);
+    return success[0].value;
+  }).then(success => {
+    if(orderType === 5) { return success.hour; }
+    else if(orderType === 4) { return success.day; }
+    else if(orderType === 2) { return success.week; }
+    else if(orderType === 3) { return success.month; }
+    else if(orderType === 6) { return success.season; }
+    else if(orderType === 7) { return success.year; }    
+  });
   const qrCode = await alipay_f2f.createQRPay({
     tradeNo: orderId,
-    subject: 'ss续费',
+    subject: orderSetting.orderName || 'ss续费',
     totalAmount: +amount,
     body: 'ss',
     timeExpress: 10,
@@ -61,6 +77,29 @@ const createOrder = async (user, account, amount, orderType = 3) => {
     orderId,
     qrCode: qrCode.qr_code,
   };
+};
+
+const sendSuccessMail = async userId => {
+  const emailPlugin = appRequire('plugins/email/index');
+  const user = await knex('user').select().where({
+    type: 'normal',
+    id: userId,
+  }).then(success => {
+    if(success.length) {
+      return success[0];
+    }
+    return Promise.reject('user not found');
+  });
+  const orderMail = await knex('webguiSetting').select().where({
+    key: 'mail',
+  }).then(success => {
+    if(!success.length) {
+      return Promise.reject('settings not found');
+    }
+    success[0].value = JSON.parse(success[0].value);
+    return success[0].value.order;
+  });
+  await emailPlugin.sendMail(user.email, orderMail.title, orderMail.content);
 };
 
 cron.minute(async () => {
@@ -93,6 +132,7 @@ cron.minute(async () => {
         });
       }).then(() => {
         logger.info(`订单支付成功: [${ order.orderId }][${ order.amount }][account: ${ accountId }]`);
+        sendSuccessMail(userId);
       }).catch(err => {
         logger.error(`订单支付失败: [${ order.orderId }]`, err);
       });
@@ -205,8 +245,9 @@ const orderListAndPaging = async (options = {}) => {
 };
 
 cron.minute(() => {
+  if(!alipay_f2f) { return; }
   knex('alipay').delete().where({ status: 'CREATE' }).whereBetween('createTime', [0, Date.now() - 1 * 24 * 3600 * 1000]).then();
-}, 30);
+}, 37);
 
 exports.orderListAndPaging = orderListAndPaging;
 exports.orderList = orderList;
